+++ app/modules/ibridge/scripts/uninstall.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+# Remove ibridge script
+rm -f "${MUNKIPATH}preflight.d/ibridge.py"
+
+# Remove ibridge.plist file
+rm -f "${CACHEPATH}ibridge.plist"