+++ app/modules/ibridge/views/ibridge_listing.php
@@ -3,14 +3,14 @@
 <?php //Initialize models needed for the table
 new Machine_model;
 new Reportdata_model;
-new Fonts_model;
+new Ibridge_model;
 ?>
 
 <div class="container">
   <div class="row">
 	<div class="col-lg-12">
 
-	  <h3><span data-i18n="fonts.reporttitle"></span> <span id="total-count" class='label label-primary'>…</span></h3>
+	  <h3><span data-i18n="ibridge.reporttitle"></span> <span id="total-count" class='label label-primary'>…</span></h3>
 
 	  <table class="table table-striped table-condensed table-bordered">
 
@@ -18,22 +18,17 @@ new Fonts_model;
 		  <tr>
 			<th data-i18n="listing.computername" data-colname='machine.computer_name'></th>
 			<th data-i18n="serial" data-colname='reportdata.serial_number'></th>
-			<th data-i18n="fonts.type_name" data-colname='fonts.type_name'></th>
-			<th data-i18n="fonts.type_enabled" data-colname='fonts.type_enabled'></th>
-			<th data-i18n="fonts.valid" data-colname='fonts.valid'></th>
-			<th data-i18n="fonts.duplicate" data-colname='fonts.duplicate'></th>
-			<th data-i18n="fonts.style" data-colname='fonts.style'></th>
-			<th data-i18n="fonts.family" data-colname='fonts.family'></th>
-			<th data-i18n="fonts.type" data-colname='fonts.type'></th>
-			<th data-i18n="fonts.copy_protected" data-colname='fonts.copy_protected'></th>
-			<th data-i18n="fonts.path" data-colname='fonts.path'></th>
-			<th data-i18n="fonts.vendor" data-colname='fonts.vendor'></th>
+			<th data-i18n="ibridge.model_name" data-colname='ibridge.model_name'></th>
+			<th data-i18n="ibridge.model_identifier" data-colname='ibridge.model_identifier'></th>
+			<th data-i18n="ibridge.build" data-colname='ibridge.build'></th>
+			<th data-i18n="ibridge.ibridge_serial_number" data-colname='ibridge.ibridge_serial_number'></th>
+			<th data-i18n="ibridge.boot_uuid" data-colname='ibridge.boot_uuid'></th>
 		  </tr>
 		</thead>
 
 		<tbody>
 		  <tr>
-			<td data-i18n="listing.loading" colspan="12" class="dataTables_empty"></td>
+			<td data-i18n="listing.loading" colspan="7" class="dataTables_empty"></td>
 		  </tr>
 		</tbody>
 
@@ -58,7 +53,6 @@ new Fonts_model;
         var mySort = [], // Initial sort
             hideThese = [], // Hidden columns
             col = 0, // Column counter
-            runtypes = [], // Array for runtype column
             columnDefs = [{ visible: false, targets: hideThese }]; //Column Definitions
 
         $('.table th').map(function(){
@@ -81,22 +75,17 @@ new Fonts_model;
                 url: appUrl + '/datatables/data',
                 type: "POST",
                 data: function(d){
-                     d.mrColNotEmpty = "name";
-
+                    d.mrColNotEmpty = "boot_uuid";
+                    
                     // Check for column in search
                     if(d.search.value){
                         $.each(d.columns, function(index, item){
-                            if(item.name == 'fonts.' + d.search.value){
+                            if(item.name == 'ibridge.' + d.search.value){
                                 d.columns[index].search.value = '> 0';
                             }
                         });
 
-                    }
-        		    // IDK what this does
-                    if(d.search.value.match(/^\d+\.\d+(\.(\d+)?)?$/)){
-                        var search = d.search.value.split('.').map(function(x){return ('0'+x).slice(-2)}).join('');
-                        d.search.value = search;
-                    }
+                    }                 
                 }
             },
             dom: mr.dt.buttonDom,
@@ -108,28 +97,9 @@ new Fonts_model;
 	        	var name=$('td:eq(0)', nRow).html();
 	        	if(name == ''){name = "No Name"};
 	        	var sn=$('td:eq(1)', nRow).html();
-	        	var link = mr.getClientDetailLink(name, sn, '#tab_fonts-tab');
+	        	var link = mr.getClientDetailLink(name, sn, '#tab_ibridge-tab');
 	        	$('td:eq(0)', nRow).html(link);
-
-	        	var internal=$('td:eq(3)', nRow).html();
-	        	internal = internal == '1' ? i18n.t('yes') :
-	        	(internal === '0' ? i18n.t('no') : '')
-	        	$('td:eq(3)', nRow).html(internal)
-
-	        	var media=$('td:eq(4)', nRow).html();
-	        	media = media == '1' ? i18n.t('yes') :
-	        	(media === '0' ? i18n.t('no') : '')
-	        	$('td:eq(4)', nRow).html(media)
-
-	        	var media=$('td:eq(5)', nRow).html();
-	        	media = media == '1' ? i18n.t('yes') :
-	        	(media === '0' ? i18n.t('no') : '')
-	        	$('td:eq(5)', nRow).html(media)
-
-	        	var media=$('td:eq(9)', nRow).html();
-	        	media = media == '1' ? i18n.t('yes') :
-	        	(media === '0' ? i18n.t('no') : '')
-	        	$('td:eq(9)', nRow).html(media)
+                
 		    }
 	    });